feat: started manage history by include param in getAllBoards query
