+++ src/store/boards.api.ts
@@ -1,5 +1,20 @@
 import { createApi } from '@reduxjs/toolkit/query/react';
 import { apiBaseQuery } from '../utils/api.utils';
+import { type List } from './lists.api';
+import { type Task } from './tasks.api';
+
+export type HistoryEntity = Partial<List> | Partial<Task> | Partial<Board>;
+export type EntityType = 'Board' | 'List' | 'Task';
+
+export interface History {
+  id: string;
+  entityType: EntityType;
+  entityId: string;
+  boardId: string;
+  operation: string;
+  oldValue?: HistoryEntity;
+  newValue?: HistoryEntity;
+}
 
 export interface Board {
   id: string;
@@ -7,6 +22,7 @@ export interface Board {
   position?: number;
   createdAt?: string;
   updatedAt?: string;
+  history?: History[];
 }
 
 export const boardsApi = createApi({
@@ -14,8 +30,11 @@ export const boardsApi = createApi({
   baseQuery: apiBaseQuery,
   tagTypes: ['Boards'],
   endpoints: (builder) => ({
-    fetchBoards: builder.query<Board[], void>({
-      query: () => 'boards',
+    fetchBoards: builder.query<Board[], { includeHistory?: boolean }>({
+      query: ({ includeHistory }) => {
+        const params = includeHistory ? '?include=history' : '';
+        return `boards${params}`;
+      },
       providesTags: ['Boards'],
     }),
     createBoard: builder.mutation<Board, { title: string }>({
