+++ src/store/store.ts
@@ -2,21 +2,18 @@ import { configureStore } from '@reduxjs/toolkit';
 import { boardsApi } from './boards.api';
 import { listsApi } from './lists.api';
 import { tasksApi } from './tasks.api';
-import { historyApi } from './history.api';
 
 export const store = configureStore({
   reducer: {
     [boardsApi.reducerPath]: boardsApi.reducer,
     [listsApi.reducerPath]: listsApi.reducer,
     [tasksApi.reducerPath]: tasksApi.reducer,
-    [historyApi.reducerPath]: historyApi.reducer,
   },
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware().concat(
       boardsApi.middleware,
       listsApi.middleware,
-      tasksApi.middleware,
-      historyApi.middleware
+      tasksApi.middleware
     ),
 });
